small fixes, added google sign up btn
