+++ src/atoms/Divider/Divider.css
@@ -1,0 +1,10 @@
+#divider-container {
+  align-items: center;
+  display: flex;
+  justify-content: center;
+  margin: 14px 0;
+}
+
+.divider-text {
+  margin: 0 6px;
+}
