+++ src/atoms/Divider/styles.tsx
@@ -1,0 +1,11 @@
+import { styled } from '@material-ui/core/styles';
+import Divider from '@material-ui/core/Divider';
+
+export const StyledDivider = styled(Divider)({
+  width: '80px',
+});
+
+// remove
+export const RemoveThis = styled(Divider)({
+  width: '80px',
+});
