+++ src/organisms/SignInForm/SignInForm.css
@@ -1,14 +1,3 @@
-#divider-container {
-  align-items: center;
-  display: flex;
-  justify-content: center;
-  margin: 14px 0;
-}
-
-.divider-text {
-  margin: 0 6px;
-}
-
 #signin {
   display: flex;
   flex-direction: column;
